프로젝트 2차 합치기 및 css 수정 퇴사 기능 구현

+++ src/main/java/com/project/ownote/annualData/dto/AnnData.java
@@ -16,8 +16,8 @@ public class AnnData {
 @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long  anndata_id;
 
-  private int  anndata_date;
-    private int anndata_add;
+  private double  anndata_date;
+    private double anndata_add;
     private double anndata_used;
     private double anndata_extra;
     private int emp_id;
